Move check log availability into component.
